+++ ocaz-simple-browser/app/src/app/components/ThumbnailContainer.js
@@ -1,0 +1,15 @@
+import Thumbnail from "./Thumbnail";
+
+export default function ThumbnailContainer({ children }) {
+  return (
+    <div
+      style={{
+        display: "flex",
+        flexWrap: "wrap",
+        justifyContent: "space-between",
+      }}
+    >
+      {children}
+    </div>
+  );
+}
