+++ ocaz-simple-browser/app/src/app/components/VideoThumbnail.js
@@ -1,0 +1,30 @@
+export default function VideoThumbnail({ object, height }) {
+  const width = Math.floor(height * (object.video.width / object.video.height));
+  return (
+    <div style={{ border: "2px solid #333333" }}>
+      <div
+        style={{
+          position: "relative",
+          width: `${width}px`,
+          height: `${height}px`,
+        }}
+      >
+        <video
+          style={{
+            position: "absolute",
+            top: "0",
+            bottom: "0",
+            left: "0",
+            right: "0",
+            width: "100%",
+            height: "auto",
+            objectFit: "contain",
+          }}
+          src={`/api/videoDigester/object/head10mbSha1/${object.head10mbSha1}`}
+          onMouseOver={(e) => e.target.play()}
+          onMouseOut={(e) => e.target.pause()}
+        />
+      </div>
+    </div>
+  );
+}
